# Improve performance of SHOW TAG VALUES