--- a/tsdb/index/tsi1/index.go
+++ b/tsdb/index/tsi1/index.go
@@ -614,21 +614,22 @@ func (i *Index) MeasurementTagKeysByExpr(name []byte, expr influxql.Expr) (map[s
 	return fs.MeasurementTagKeysByExpr(name, expr)
 }
 
-func (i *Index) MeasurementTagKeyValuesByExpr(name, key []byte, expr influxql.Expr) (map[string]struct{}, error) {
-	fs := i.RetainFileSet()
-	defer fs.Release()
+func (i *Index) MeasurementTagKeyValuesByExpr(name []byte, key []string, expr influxql.Expr, keysSorted bool) ([][]string, error) {
+	return nil, nil
+	// fs := i.RetainFileSet()
+	// defer fs.Release()
 
-	values := make(map[string]struct{})
-	// If there is not expr, we return all tag values for the key
-	if expr == nil {
-		itr := fs.TagValueIterator(name, key)
-		for val := itr.Next(); val != nil; val = itr.Next() {
-			values[string(val.Value())] = struct{}{}
-		}
-		return values, nil
-	}
+	// values := make(map[string]struct{})
+	// // If there is not expr, we return all tag values for the key
+	// if expr == nil {
+	// 	itr := fs.TagValueIterator(name, key)
+	// 	for val := itr.Next(); val != nil; val = itr.Next() {
+	// 		values[string(val.Value())] = struct{}{}
+	// 	}
+	// 	return values, nil
+	// }
 
-	return fs.tagValuesByKeyAndExpr(name, key, expr, i.fieldset)
+	// return fs.tagValuesByKeyAndExpr(name, key, expr, i.fieldset)
 }
 
 // ForEachMeasurementSeriesByExpr iterates over all series in a measurement filtered by an expression.