###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated

This PR refactors the `tsdb.Store`'s `TagValues` method to significantly improve the performance of `SHOW TAG VALUES` from that experienced in `1.3.1`.

In some cases I've observed some queries completing around 185x faster than `1.3.1`. Fixes #8640.

It works as follows:

 - For the `inmem` index we don't need to check each shard for tag values, since all shards have a copy of the same index. Therefore we just check the first shard.
 - Instead of gathering everything up into large collections and then sorting using Quicksort, we merge multiple streams of tag values together for each shard.
 - We add new methods to the `Index` for more efficiently gathering collections of tag values for tag keys.
 - Where the `SHOW TAG VALUES` doesn't include a `WHERE` clause, we skip more expensive methods that filter series, and just grab all of the appropriate series.
 - We push all the tag keys we need values for down into the index and gather all relevant values in one go, rather than repeating the process for each tag key.
 - We avoid the use of `sort.Sort` and `maps` in general.
 - We reduce wasted allocations by delaying the allocation of `KeyValue` objects, which may end up being thrown away through de-duplication anyway.
 - We leverage the invariant that tag keys are passed to the index sorted, so we can exit early from `models.Tags` iteration.
 - Where appropriate we use unsafe `string --> []byte` and `[]byte --> string` (though I'm not sure how much we're saving given the Go compiler might have kept these conversions on the stack anyway.

#### Benchmark Results

I added some benchmark. Here are the results comparing `v1.3.1` to these changes:

```
benchmark                                                                                   old ns/op      new ns/op     delta
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=1_v=100-8           93631          24872         -73.44%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=1_v=1000-8          1058656        271110        -74.39%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=10_v=100-8          1011901        242419        -76.04%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=10_v=1000-8         11425817       2741175       -76.01%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=100_v=100-8         10045286       2506089       -75.05%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=100_v=1000-8        121460151      29636546      -75.60%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=1_v=100-8          5535025        27494         -99.50%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=1_v=1000-8         57200134       277599        -99.51%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=10_v=100-8         61137967       264608        -99.57%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=10_v=1000-8        716986198      2758907       -99.62%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=100_v=100-8        644121420      2685845       -99.58%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=100_v=1000-8       7525538172     26869192      -99.64%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=1_v=100-8         46656          33858         -27.43%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=1_v=1000-8        466128         319074        -31.55%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=10_v=100-8        554105         361193        -34.82%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=10_v=1000-8       5748353        3790884       -34.05%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=100_v=100-8       5430254        3673582       -32.35%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=100_v=1000-8      61151803       40770214      -33.33%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=1_v=100-8        2968088        221329        -92.54%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=1_v=1000-8       29958242       2263208       -92.45%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=10_v=100-8       31144989       2104543       -93.24%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=10_v=1000-8      380440093      26534566      -93.03%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=100_v=100-8      359041949      24246682      -93.25%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=100_v=1000-8     4443898489     312383241     -92.97%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=1_v=100-8            92736          23104         -75.09%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=1_v=1000-8           1038086        257750        -75.17%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=10_v=100-8           1005688        227052        -77.42%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=10_v=1000-8          11498716       2689217       -76.61%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=100_v=100-8          10179007       2460154       -75.83%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=100_v=1000-8         119252644      29451575      -75.30%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=1_v=100-8           6282477        276993        -95.59%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=1_v=1000-8          65477559       3715131       -94.33%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=10_v=100-8          69200875       2740942       -96.04%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=10_v=1000-8         866964826      41253818      -95.24%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=100_v=100-8         754756036      30240493      -95.99%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=100_v=1000-8        8933943418     383221060     -95.71%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=1_v=100-8          39095          28946         -25.96%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=1_v=1000-8         527935         370338        -29.85%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=10_v=100-8         520686         337172        -35.24%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=10_v=1000-8        5671519        3657734       -35.51%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=100_v=100-8        5469111        3677777       -32.75%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=100_v=1000-8       61889900       41131446      -33.54%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=10_m=1_v=100-8         3275654        397360        -87.87%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=10_m=1_v=1000-8        32386777       4023906       -87.58%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=10_m=10_v=100-8        35536831       3915073       -88.98%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=10_m=10_v=1000-8       445130899      50358133      -88.69%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=10_m=100_v=100-8       410576229      43287794      -89.46%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=10_m=100_v=1000-8      4805356459     486141537     -89.88%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=1_m=1_v=100-8            114199         46340         -59.42%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=1_m=1_v=1000-8           1402464        470217        -66.47%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=1_m=10_v=100-8           1217745        463179        -61.96%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=1_m=10_v=1000-8          15318214       5046922       -67.05%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=1_m=100_v=100-8          12327456       4974928       -59.64%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=1_m=100_v=1000-8         167473481      54547135      -67.43%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=10_m=1_v=100-8           876073         472860        -46.03%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=10_m=1_v=1000-8          10026211       4810940       -52.02%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=10_m=10_v=100-8          9319077        4726873       -49.28%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=10_m=10_v=1000-8         118900081      50801562      -57.27%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=10_m=100_v=100-8         103685056      50633264      -51.17%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Filtered_s=10_m=100_v=1000-8        1218726014     502119575     -58.80%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=1_m=1_v=100-8          54435          46743         -14.13%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=1_m=1_v=1000-8         680329         551619        -18.92%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=1_m=10_v=100-8         660864         515792        -21.95%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=1_m=10_v=1000-8        7326683        5613384       -23.38%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=1_m=100_v=100-8        6432336        5149298       -19.95%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=1_m=100_v=1000-8       80600653       66508896      -17.48%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=10_m=1_v=100-8         482016         540514        +12.14%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=10_m=1_v=1000-8        5063378        5762606       +13.81%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=10_m=10_v=100-8        5164921        5375245       +4.07%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=10_m=10_v=1000-8       59156848       62785660      +6.13%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=10_m=100_v=100-8       58329304       59382389      +1.81%
BenchmarkStore_TagValues/random_values=false_index=tsi1_Unfiltered_s=10_m=100_v=1000-8      643895491      647552903     +0.57%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=1_m=1_v=100-8             115727         46405         -59.90%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=1_m=1_v=1000-8            1440140        465606        -67.67%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=1_m=10_v=100-8            1234813        459918        -62.75%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=1_m=10_v=1000-8           15755860       4956455       -68.54%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=1_m=100_v=100-8           12596113       4898969       -61.11%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=1_m=100_v=1000-8          105677861      21834289      -79.34%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=10_m=1_v=100-8            1354048        726686        -46.33%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=10_m=1_v=1000-8           15317751       7584517       -50.49%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=10_m=10_v=100-8           14247949       7283196       -48.88%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=10_m=10_v=1000-8          176247072      81680375      -53.66%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=10_m=100_v=100-8          159042762      77424237      -51.32%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Filtered_s=10_m=100_v=1000-8         1357106565     453632910     -66.57%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=1_m=1_v=100-8           56688          47943         -15.43%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=1_m=1_v=1000-8          671970         539788        -19.67%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=1_m=10_v=100-8          656781         497756        -24.21%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=1_m=10_v=1000-8         7462946        5704081       -23.57%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=1_m=100_v=100-8         6674778        5291050       -20.73%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=1_m=100_v=1000-8        61318155       35040917      -42.85%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=10_m=1_v=100-8          657046         603340        -8.17%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=10_m=1_v=1000-8         7303178        7099275       -2.79%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=10_m=10_v=100-8         7221825        6556733       -9.21%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=10_m=10_v=1000-8        86004814       81364506      -5.40%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=10_m=100_v=100-8        82342911       74856378      -9.09%
BenchmarkStore_TagValues/random_values=true_index=tsi1_Unfiltered_s=10_m=100_v=1000-8       593101145      505786539     -14.72%

benchmark                                                                                   old allocs     new allocs     delta
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=1_v=100-8           451            49             -89.14%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=1_v=1000-8          4107           52             -98.73%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=10_v=100-8          4294           283            -93.41%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=10_v=1000-8         40852          313            -99.23%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=100_v=100-8         42687          2623           -93.86%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=1_m=100_v=1000-8        408239         2923           -99.28%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=1_v=100-8          40192          53             -99.87%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=1_v=1000-8         400243         56             -99.99%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=10_v=100-8         401420         287            -99.93%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=10_v=1000-8        4001956        317            -99.99%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=100_v=100-8        4013666        2627           -99.93%
BenchmarkStore_TagValues/random_values=false_index=inmem_Filtered_s=10_m=100_v=1000-8       40019044       2927           -99.99%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=1_v=100-8         259            172            -33.59%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=1_v=1000-8        1880           999            -46.86%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=10_v=100-8        2347           1432           -38.99%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=10_v=1000-8       21129          11008          -47.90%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=100_v=100-8       22363          13609          -39.15%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=1_m=100_v=1000-8      204174         106323         -47.93%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=1_v=100-8        21088          1128           -94.65%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=1_v=1000-8       204021         10330          -94.94%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=10_v=100-8       198317         10300          -94.81%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=10_v=1000-8      2003092        101100         -94.95%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=100_v=100-8      2016907        104341         -94.83%
BenchmarkStore_TagValues/random_values=false_index=inmem_Unfiltered_s=10_m=100_v=1000-8     20018852       1010003        -94.95%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=1_v=100-8            451            49             -89.14%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=1_v=1000-8           4095           52             -98.73%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=10_v=100-8           4294           283            -93.41%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=10_v=1000-8          40735          313            -99.23%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=100_v=100-8          42675          2623           -93.85%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=1_m=100_v=1000-8         407249         2923           -99.28%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=1_v=100-8           40245          56             -99.86%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=1_v=1000-8          399516         65             -99.98%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=10_v=100-8          401843         317            -99.92%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=10_v=1000-8         3997028        407            -99.99%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=100_v=100-8         4017980        2927           -99.93%
BenchmarkStore_TagValues/random_values=true_index=inmem_Filtered_s=10_m=100_v=1000-8        39966776       3827           -99.99%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=1_v=100-8          214            149            -30.37%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=1_v=1000-8         2145           1134           -47.13%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=10_v=100-8         2222           1368           -38.43%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=10_v=1000-8        20402          10639          -47.85%
BenchmarkStore_TagValues/random_values=true_index=inmem_Unfiltered_s=1_m=100_v=100-8        22052          13457          -38.98%
…